New Updates 2
Fixed the probelm of Gamplay 3rd round could not be process to the EndofGameSettlement Problem: The See Result Button now is able to be clicked after the newly set up of checking the Player 2's state of Offer Accept or the trigger of Offer Cancelled during Negotiation with ArtCollector.    Prompts are now updates for better diversity of artwork image generation.